chore(breadcrumb): rename breadcrumb item file
