+++ packages/react-magma-dom/src/components/Breadcrumb/Breadcrumb.stories.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import { Breadcrumb } from '.';
-import { BreadcrumbItem } from './Item';
+import { Breadcrumb, BreadcrumbItem } from '.';
 
 export default {
   component: Breadcrumb,
