+++ packages/react-magma-dom/src/components/Breadcrumb/index.tsx
@@ -1,2 +1,2 @@
 export * from './Breadcrumb';
-export * from './Item';
+export * from './BreadcrumbItem';
